Update ephemeral media viewer

+++ lib/features/chat/widgets/ephemeral_media_viewer_screen.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 import 'dart:typed_data';
+import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter/material.dart';
 import 'package:path_provider/path_provider.dart';
 import 'package:video_player/video_player.dart';
@@ -95,12 +96,17 @@ class _EphemeralMediaViewerScreenState extends State<EphemeralMediaViewerScreen>
 
   Future<void> _initVideo() async {
     try {
-      final tempDir = await getTemporaryDirectory();
-      final tempPath = '${tempDir.path}/ephemeral_${DateTime.now().millisecondsSinceEpoch}.mp4';
-      _tempVideoFile = File(tempPath);
-      await _tempVideoFile!.writeAsBytes(widget.rawBytes, flush: true);
+      if (kIsWeb) {
+        final uri = Uri.dataFromBytes(widget.rawBytes, mimeType: 'video/mp4');
+        _videoController = VideoPlayerController.networkUrl(uri);
+      } else {
+        final tempDir = await getTemporaryDirectory();
+        final tempPath = '${tempDir.path}/ephemeral_${DateTime.now().millisecondsSinceEpoch}.mp4';
+        _tempVideoFile = File(tempPath);
+        await _tempVideoFile!.writeAsBytes(widget.rawBytes, flush: true);
+        _videoController = VideoPlayerController.file(_tempVideoFile!);
+      }
 
-      _videoController = VideoPlayerController.file(_tempVideoFile!);
       await _videoController!.initialize();
       _videoController!.addListener(_onVideoUpdate);
       if (mounted) {
@@ -110,7 +116,9 @@ class _EphemeralMediaViewerScreenState extends State<EphemeralMediaViewerScreen>
         _videoController!.setLooping(false);
         _videoController!.play();
       }
-    } catch (_) {}
+    } catch (e) {
+      debugPrint('[EphemeralMediaViewer] Error initializing video: $e');
+    }
   }
 
   @override
@@ -121,7 +129,7 @@ class _EphemeralMediaViewerScreenState extends State<EphemeralMediaViewerScreen>
     _zoomAnimController.dispose();
     _videoController?.removeListener(_onVideoUpdate);
     _videoController?.dispose();
-    if (_tempVideoFile != null && _tempVideoFile!.existsSync()) {
+    if (!kIsWeb && _tempVideoFile != null && _tempVideoFile!.existsSync()) {
       try {
         _tempVideoFile!.deleteSync();
       } catch (_) {}
